requirements.txt
coloredlogs==15.0.1
gradio==4.7.1
openai==0.27.8
Pillow==9.5.0
Pillow==10.1.0
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
